requirements.txt
setuptools>=69.0.0
undetected-chromedriver>=3.5.5
selenium>=4.15.0
requests>=2.31.0
psutil>=5.9.0

# Cloud Storage (opsional — simpan link ke Google Sheets)
gspread>=6.0.0
google-auth>=2.25.0
